Allow prterun to act as prun
Provide a "--dvm[=<arg>]" cmd line directive telling prterun
to use an existing persistent DVM instead of instantiating its
own one-shot DVM. Move the body of prun into libprrte so that
both prun and prterun can utilize it without code duplication.

Have prterun check for the "--dvm" directive as soon as it has
set the personality and checked for proxy operation. If found,
translate any included options into the appropriate directives
understood by prun and call the common prun code.

Update both ompi and prte schizo components to include the
new option. Update both ompi and prte schizo help files to
support the new option.

(cherry picked from commit e2854d55d197d59c515d5df1ad6ed040fc6fb000)
