+++ src/mca/schizo/ompi/schizo_ompi.c
@@ -97,6 +97,7 @@ static struct option ompioptions[] = {
     PMIX_OPTION_SHORT_DEFINE(PRTE_CLI_PARSEABLE, PMIX_ARG_NONE, 'p'),
     PMIX_OPTION_SHORT_DEFINE(PRTE_CLI_PARSABLE, PMIX_ARG_NONE, 'p'), // synonym for parseable
     PMIX_OPTION_DEFINE(PRTE_CLI_PERSONALITY, PMIX_ARG_REQD),
+    PMIX_OPTION_DEFINE(PRTE_CLI_DVM, PMIX_ARG_OPTIONAL),
 
     // MCA parameters
     PMIX_OPTION_DEFINE(PRTE_CLI_PRTEMCA, PMIX_ARG_REQD),
