+++ src/prted/pmix/pmix_server.c
@@ -695,7 +695,6 @@ int pmix_server_init(void)
         }
     }
 
-#ifdef PMIX_SINGLETON
     /* if we were started to support a singleton, then let the server library know */
     if (NULL != prte_pmix_server_globals.singleton) {
         PMIX_INFO_LIST_ADD(prc, ilist, PMIX_SINGLETON,
@@ -706,16 +705,13 @@ int pmix_server_init(void)
             return rc;
         }
     }
-#endif
 
-#ifdef PMIX_BIND_PROGRESS_THREAD
     if (NULL != prte_progress_thread_cpus) {
         PMIX_INFO_LIST_ADD(prc, ilist, PMIX_BIND_PROGRESS_THREAD,
                            prte_progress_thread_cpus, PMIX_STRING);
         PMIX_INFO_LIST_ADD(prc, ilist, PMIX_BIND_REQUIRED,
                            &prte_bind_progress_thread_reqd, PMIX_BOOL);
     }
-#endif
 
     /* if we are the MASTER, then we are the scheduler
      * as well as a gateway */
